requirements.txt
Flask
openai
python-dotenv
pymongo
google-api-python-client
google-auth
google-auth-oauthlib
apscheduler
twilio
deep-translator
langdetect
pandas
requests
cryptography
requests
pillow
uuid
markdown
cloudinary
redis
gunicorn
sentry-sdk
rq
flask_cors
